define project build requirements
this will make sure that pip install the listed
packages when building the project.

Because we are building a cython extension and we have to
link against numpy headers the setup.py file included some checks
to make sure Cython was available and deferred getting the
numpy headers till building the extension.

This can be done by specifying the packages required in the
build environment by relying on the PEP 517 and 518.

As of PEP 518 we have now made sure that when pip creates
a virtual environment for the build it will add setuptools,
wheel, cython and numpy before starting building.

We can then take the checks away and be sure that the
building environment is reproducible.